+++ factors.h
@@ -2,9 +2,9 @@
 #define FACTORS_H
 #include <stdio.h>
 
-int square_root(long long int n);
-void factorize(long long int num);
-int isPrime(long long num);
-void findPrimeFactors(long long n);
+int square_root(long int n);
+void factorize(long int num);
+int isPrime(long int num);
+void findPrimeFactors(long int n);
 
 #endif /* FACTORS_H */
